+++ example/lib/pages/highlights.page.dart
@@ -7,7 +7,7 @@ import 'package:visual_editor/shared/utils/string.utils.dart';
 import 'package:visual_editor/visual-editor.dart';
 
 import '../const/sample-highlights.const.dart';
-import '../widgets/demo-scaffold.dart';
+import '../widgets/demo-page-scaffold.dart';
 import '../widgets/loading.dart';
 import '../widgets/selection-stats.dart';
 
@@ -51,7 +51,7 @@ class _HighlightsPageState extends State<HighlightsPage> {
               ],
       );
 
-  Widget _scaffold({required List<Widget> children}) => DemoScaffold(
+  Widget _scaffold({required List<Widget> children}) => DemoPageScaffold(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: children,