+++ example/lib/pages/markers-attachments.page.dart
@@ -7,7 +7,7 @@ import 'package:visual_editor/visual-editor.dart';
 
 import '../const/dimensions.const.dart';
 import '../models/markers-and-scroll-offset.model.dart';
-import '../widgets/demo-scaffold.dart';
+import '../widgets/demo-page-scaffold.dart';
 import '../widgets/loading.dart';
 import '../widgets/markers-attachments.dart';
 
@@ -64,7 +64,7 @@ class _MarkersAttachmentsPageState extends State<MarkersAttachmentsPage> {
               ],
       );
 
-  Widget _scaffold({required List<Widget> children}) => DemoScaffold(
+  Widget _scaffold({required List<Widget> children}) => DemoPageScaffold(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: children,